requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.28.2
beautifulsoup4==4.12.2
transformers==4.30.2
torch==2.0.1
pillow==9.5.0
reportlab==4.0.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
marked==0.1.5
htmx.org==1.9.6
